chore: optimize markdown anchors
