+++ src/lib/replaceInternalHref.js
@@ -1,4 +1,4 @@
-import { navigate } from 'gatsby-plugin-react-intl'
+import { navigateInsideEventListener } from './utils'
 import { withPrefix } from 'gatsby'
 
 const reAnchor = /[^-\w\u4E00-\u9FFF]*/g // with CJKLanguage
@@ -24,11 +24,7 @@ export default function replaceInternalHref(doc, version, simpletab = false) {
       const hrefArray = href.split('/')
       href = hrefArray[hrefArray.length - 1].replace('.md', '')
       a.href = withPrefix([doc, version, href].join('/'))
-      a.addEventListener('click', e => {
-        e.preventDefault()
-
-        navigate(e.target.getAttribute('href'))
-      })
+      a.addEventListener('click', navigateInsideEventListener)
     }
 
     if (a.classList.contains('anchor')) {
